Prefer other types to ‘int’
* src/merge.c (merge_result, merge_hunk):
Hunk number is intmax_t, not int, fixing an unlikely overflow.
* src/patch.c (invc): Now signed char, not int.
(numeric_string): Use bool, not int, for sign.
* src/pch.c (p_says_nonexistent, sha1_says_nonexistent)
(pch_says_nonexistent): Now char, not int.  All uses changed.
(p_rfc834_nesting, pget_line): Use idx_t, not int,
fixing an unlikely overflow.  All uses changed.
(another_hunk): Now signed char, not int.  All uses changed.
* src/util.c (format_linenum): Remove unnecessary casts to int.

+++ src/merge.c
@@ -165,13 +165,14 @@ print_linerange (lin from, lin to)
 }
 
 static void
-merge_result (bool *first_result, int hunk, char const *what, lin from, lin to)
+merge_result (bool *first_result, intmax_t hunk, char const *what,
+	      lin from, lin to)
 {
   static char const *last_what;
 
   if (*first_result && what)
     {
-      printf ("Hunk #%d %s at ", hunk, what);
+      printf ("Hunk #%jd %s at ", hunk, what);
       last_what = what;
     }
   else if (! what)
@@ -193,7 +194,8 @@ merge_result (bool *first_result, int hunk, char const *what, lin from, lin to)
 }
 
 bool
-merge_hunk (int hunk, struct outstate *outstate, lin where, bool *somefailed)
+merge_hunk (intmax_t hunk, struct outstate *outstate,
+	    lin where, bool *somefailed)
 {
   bool applies_cleanly;
   bool first_result = true;
